rename and move LexboxApiHub.cs to CrdtMiniLcmApiHub.cs
