+++ backend/LocalWebApp/Routes/ActivityRoutes.cs
@@ -1,6 +1,7 @@
 ﻿using Crdt.Changes;
 using Crdt.Core;
 using Crdt.Db;
+using LocalWebApp.Hubs;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.OpenApi.Models;
 
@@ -14,7 +15,7 @@ public static class ActivityRoutes
         {
             operation.Parameters.Add(new OpenApiParameter()
             {
-                Name = LexboxApiHub.ProjectRouteKey,
+                Name = CrdtMiniLcmApiHub.ProjectRouteKey,
                 In = ParameterLocation.Path,
                 Required = true
             });
